requirements_hydra.txt
# Dependencias para Hydra
hydra-core>=1.3.0
omegaconf>=2.3.0

# Dependencias existentes del proyecto (mantener las versiones actuales)
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=0.11.0
wandb>=0.15.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0

# Dependencias adicionales para notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0 